% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_family_structure_data}
\alias{create_family_structure_data}
\title{Create a random dataset of families}
\usage{
create_family_structure_data(
  seed = -1,
  n = 1000,
  adult_index_case_prob = 0.8,
  T_max = 21
)
}
\arguments{
\item{seed}{seed for the simulation}

\item{n}{number of families}

\item{adult_index_case_prob}{probability that the index case is adult}

\item{T_max}{number of days to run epidemic (scalar or vector of size n)}
}
\value{
a the result as a tibble including the number of adults, the number of children, the index case, and the epidemic duration (in days) of each family
}
\description{
Create a random dataset of families
}
\seealso{
\code{create_simulated_data} \code{run_simulation}
}
