% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_simulated_data}
\alias{create_simulated_data}
\title{Simulate the epidemic with in a household}
\usage{
create_simulated_data(
  seed = -1,
  family_structure_data,
  beta_cc,
  beta_ca,
  beta_ac,
  beta_aa,
  P = NULL,
  T_max = 45,
  ssn = 1
)
}
\arguments{
\item{seed}{seed for the simulation}

\item{family_structure_data}{dataset of families including the number of adults, number of children, and the index case (see \code{create_family_structure_data})}

\item{beta_cc}{proabilty that a child will infect a child}

\item{beta_ca}{proabilty that an adult will infect a child}

\item{beta_ac}{proabilty that a child will infect an adult}

\item{beta_aa}{proabilty that an adult will infect an adult}

\item{P}{a vector of generation time distribution (default is generated by \code{load_P})}

\item{T_max}{the length of the epidemic within the family}

\item{ssn}{the number of households to simulate}
}
\value{
a tibble with the epidemic development over time by day
}
\description{
Simulate the epidemic with in a household
}
\seealso{
\code{create_family_structure_data} \code{load_P}
}
