% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{load_P}
\alias{load_P}
\title{Generate a vector of the generation time distribution using the Gamma distribution}
\usage{
load_P(mean_P = 4.5, sd_P = 2.5)
}
\arguments{
\item{mean_P}{the mean of the distribution}

\item{sd_P}{the standard deviation of the distribution}
}
\value{
a vector of the generation time distribution
}
\description{
Generate a vector of the generation time distribution using the Gamma distribution
}
