% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulate nsim simulations of an epidemic on a dataset of households}
\usage{
run_simulation(
  nsim = 1,
  family_structure_data,
  beta_aa,
  gamma_val,
  delta_val,
  P = NULL,
  T_max = 45,
  future_plan = "sequential",
  dictionary_path = NULL
)
}
\arguments{
\item{nsim}{number of simulation to run}

\item{family_structure_data}{dataset of families including the number of adults, number of children, and the index case (see \code{create_family_structure_data})}

\item{beta_aa}{proabilty that an adult will infect an adult (must be appear in the dictionary)}

\item{gamma_val}{relative susceptibility of children to adults (must be appear in the dictionary)}

\item{delta_val}{relative infectivity of children to adults (must be appear in the dictionary)}

\item{P}{a vector of generation time distribution (default is generated by `load_P``)}

\item{T_max}{the length of the epidemic within the family}

\item{future_plan}{the plan for using the future package (default is \code{sequential})}

\item{dictionary_path}{path to the folder that contain the dictionaries}
}
\value{
a tibble with the maximum likelihood for each simulated dataset
}
\description{
Simulate nsim simulations of an epidemic on a dataset of households
}
\seealso{
\code{load_P} \code{create_family_structure_data}
}
